[Update] ArcV is now (finally) built as a library instead of an exec! [Add] ArcV is now (finally) building examples along itself!

+++ src/ArcV/Processing/Sobel.cpp
@@ -17,10 +17,14 @@ Sobel::Sobel(const Matrix<>& mat) {
 }
 
 Matrix<> Sobel::computeGradientDirection() const {
-  Matrix<> res(sobelMat.getWidth(), sobelMat.getHeight(), sobelMat.getChannelCount(), sobelMat.getImgBitDepth(), sobelMat.getColorspace());
+  Matrix<> res(sobelMat.getWidth(),
+               sobelMat.getHeight(),
+               sobelMat.getChannelCount(),
+               sobelMat.getImgBitDepth(),
+               sobelMat.getColorspace());
 
   for (std::size_t i = 0; i < sobelMat.getData().size(); ++i)
-    res[i] = 180.f + std::atan2(verticalGradient[i], horizontalGradient[i]) * (180.f / M_PI);
+    res[i] = 180.f + std::atan2(verticalGradient[i], horizontalGradient[i]) * (180.f / static_cast<float>(M_PI));
 
   return res;
 }
